+++ app/src/js/components/Executions/execution-status.js
@@ -3,19 +3,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import get from 'lodash.get';
-import cloneDeep from 'lodash.clonedeep';
 import { getExecutionStatus, getCumulusInstanceMetadata } from '../../actions';
 import { displayCase, fullDate, parseJson } from '../../utils/format';
 import { withRouter, Link } from 'react-router-dom';
 import { kibanaExecutionLink } from '../../utils/kibana';
-
-import { tableColumns } from '../../utils/table-config/execution-status';
+import { window } from '../../utils/browser';
 
 import ErrorReport from '../Errors/report';
 
 import ExecutionStatusGraph from './execution-status-graph';
-import { getEventDetails } from './execution-graph-utils';
-import SortableTable from '../SortableTable/SortableTable';
 import Metadata from '../Table/Metadata';
 import DefaultModal from '../Modal/modal';
 
@@ -24,7 +20,6 @@ class ExecutionStatus extends React.Component {
     super();
     this.navigateBack = this.navigateBack.bind(this);
     this.errors = this.errors.bind(this);
-    this.renderEvents = this.renderEvents.bind(this);
     this.openModal = this.openModal.bind(this);
     this.closeModal = this.closeModal.bind(this);
     this.state = {
@@ -69,32 +64,14 @@ class ExecutionStatus extends React.Component {
     }
   }
 
-  renderEvents () {
-    const { executionStatus } = this.props;
-    const { executionHistory: { events } } = executionStatus;
-    const mutableEvents = cloneDeep(events);
-    mutableEvents.forEach((event) => {
-      event.eventDetails = getEventDetails(event);
-    });
-
-    return (
-      <SortableTable
-        data={mutableEvents.sort((a, b) => a.id > b.id ? 1 : -1)}
-        dispatch={this.props.dispatch}
-        tableColumns={tableColumns}
-        rowId='id'
-        sortIdx='id'
-        props={[]}
-        order='asc'
-      />
-    );
-  }
-
   render () {
     const { showInputModal, showOutputModal } = this.state;
     const { executionStatus, cumulusInstance } = this.props;
-    const { execution } = executionStatus;
+    const { execution, executionHistory, stateMachine } = executionStatus;
+
     if (!execution) return null;
+
+    const { name } = execution;
 
     const errors = this.errors();
 
@@ -102,7 +79,11 @@ class ExecutionStatus extends React.Component {
       {
         label: 'Execution Status',
         property: 'status',
-        accessor: displayCase
+        accessor: d => {
+          return (
+            <span className={`execution-status-${d.toLowerCase()}`}>{displayCase(d)}</span>
+          );
+        }
       },
       {
         label: 'Execution Arn',
@@ -116,6 +97,7 @@ class ExecutionStatus extends React.Component {
         label: 'Async Operation ID',
         property: 'output',
         accessor: d => {
+          if (!d) return;
           const outputJson = JSON.parse(d);
           return get(outputJson.cumulus_meta, 'asyncOperationId');
         }
@@ -131,7 +113,7 @@ class ExecutionStatus extends React.Component {
         accessor: fullDate
       },
       {
-        label: 'Parent Workflow Exectuion',
+        label: 'Parent Workflow Execution',
         property: 'input',
         accessor: d => {
           if (!d) return 'N/A';
@@ -153,7 +135,7 @@ class ExecutionStatus extends React.Component {
               <>
                 <button
                   onClick={() => this.openModal('input')}
-                  className='button button--small'
+                  className='button button--small button--no-icon'
                 >Show Input</button>
                 <DefaultModal
                   showModal={showInputModal}
@@ -162,7 +144,7 @@ class ExecutionStatus extends React.Component {
                   hasConfirmButton={false}
                   cancelButtonClass='button--close'
                   cancelButtonText='Close'
-                  className='execution--modal'
+                  className='execution__modal'
                 >
                   <pre>{parseJson(d)}</pre>
                 </DefaultModal>
@@ -179,26 +161,31 @@ class ExecutionStatus extends React.Component {
         accessor: d => {
           if (d) {
             const jsonData = new Blob([d], { type: 'text/json' });
+            const downloadUrl = window ? window.URL.createObjectURL(jsonData) : '';
             return (
               <>
                 <button
                   onClick={() => this.openModal('output')}
-                  className='button button--small'
+                  className='button button--small button--no-icon'
                 >Show Output</button>
                 <DefaultModal
                   showModal={showOutputModal}
-                  title='Execution Output'
+                  title={
+                    <>
+                      <span>Execution Output</span>
+                      <a className='button button--small button--download button--green form-group__element--right'
+                        id='download_link'
+                        download='output.json'
+                        href={downloadUrl}
+                      >Download File</a>
+                    </>
+                  }
                   onCloseModal={() => this.closeModal('output')}
                   hasConfirmButton={false}
                   cancelButtonClass='button--close'
                   cancelButtonText='Close'
-                  className='execution--modal'
+                  className='execution__modal'
                 >
-                  <a className='button button--small button--download button--green'
-                    id='download_link'
-                    download='output.json'
-                    href={window.URL.createObjectURL(jsonData)}
-                  >Download File</a>
                   <pre>{parseJson(d)}</pre>
                 </DefaultModal>
               </>
@@ -210,13 +197,14 @@ class ExecutionStatus extends React.Component {
       },
       {
         label: 'Logs',
-        property: 'name',
+        property: 'executionArn',
         accessor: d => {
           const kibanaLink = kibanaExecutionLink(cumulusInstance, d);
+          const className = 'button button--small button__goto button__arrow button__animation button__arrow--white';
           if (kibanaLink && kibanaLink.length) {
-            return <a href={kibanaLink} target="_blank">View Logs in Kibana</a>;
+            return <a href={kibanaLink} target="_blank" className={className}>View Logs in Kibana</a>;
           } else {
-            return <Link to={'/executions/execution/' + d + '/logs'} title={d + '/logs'}>View Execution Logs</Link>;
+            return <Link to={'/executions/execution/' + d + '/logs'} title={d + '/logs'} className={className}>View Execution Logs</Link>;
           }
         }
       }
@@ -226,47 +214,30 @@ class ExecutionStatus extends React.Component {
       <div className='page__component'>
         <section className='page__section page__section__header-wrapper'>
           <h1 className='heading--large heading--shared-content with-description width--three-quarters'>
-            Execution {executionStatus.arn}
+            Execution {name}
           </h1>
 
-          {errors.length ? <ErrorReport report={errors} /> : null}
+          {(errors.length > 0) && <ErrorReport report={errors} />}
         </section>
 
         {/* stateMachine's definition and executionHistory's event statuses are needed to draw the graph */}
-        {
-          (executionStatus.stateMachine && executionStatus.executionHistory)
-            ? <section className='page__section'>
-              <div className='heading__wrapper--border'>
-                <h2 className='heading--medium with-description'>Visual workflow</h2>
-              </div>
-
-              <ExecutionStatusGraph executionStatus={executionStatus} />
-            </section>
-            : null
+        {(stateMachine && executionHistory) &&
+          <section className='page__section'>
+            <div className='heading__wrapper--border'>
+              <h2 className='heading--medium with-description'>Visual</h2>
+            </div>
+            <ExecutionStatusGraph executionStatus={executionStatus} />
+          </section>
         }
 
         <section className='page__section'>
           <div className='heading__wrapper--border'>
-            <h2 className='heading--medium with-description'>Execution Details</h2>
+            <h2 className='heading--medium with-description'>Details</h2>
           </div>
           <div className='execution__content status--process'>
-            <Metadata data={executionStatus.execution} accessors={metaAccessors} />
+            <Metadata data={execution} accessors={metaAccessors} />
           </div>
         </section>
-
-        {(executionStatus.executionHistory)
-          ? <section className='page__section'>
-            <div className='heading__wrapper--border'>
-              <h2 className='heading--medium with-description'>Events</h2>
-              <p>To find all task name and versions, select <b>More details</b> for the last Lambda- or Activity-type event. There you should find a key / value pair "workflow_tasks" which lists all tasks' version, name and arn.</p>
-              <p><b>NOTE:</b>Task / version tracking is enabled as of Cumulus version 1.9.1.</p>
-              <p><b>NOTE:</b>If the task output is greater than 10KB, the full message will be stored in an S3 Bucket. In these scenarios, task and version numbers are not part of the Lambda or Activity event output.</p>
-            </div>
-
-            {this.renderEvents()}
-          </section>
-          : null}
-
       </div>
     );
   }